CA-210925: front-to-back device links under /xenops
The problem fixed by this commit:
When shutting down a guest, we were deriving the list of devices
to be cleaned up from the values of the guest domain's nodes at
/local/domain/N/device/.../backend in xenstore; these may have
been updated by the guest, in which case the shutdown could leave
the guest domain in a bad state.

The fix:
Now when we add a device, as well as writing to
/local/domain/N/device/.../backend
we also write the same to
/xenops/domain/N/device/.../backend
and then refer to this reliable data where appropriate:
the whole /xenops tree is accessible by dom0 only.

(This is the second commit of a series of four; also it depends
on the mkdirperms function being provided by the Xenstore module.)

(cherry picked from commit 2cbbb819adc255d285604795a267a6844170cc9e)
(...which is in the xenopsd repository.)

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>

Conflicts:

	ocaml/xenops/device.ml
	ocaml/xenops/device_common.ml
	ocaml/xenops/domain.ml
	xc/device_common.mli

(cherry picked from commit d3160b3a3ec0a913d111273e69da048cfe55cef4)
Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>
(cherry picked from commit 75a9a86e4e57b20040b597eb1beda850d91b9cb8)

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>